refactor api (#8)
