+++ src/app.py
@@ -1,13 +1,16 @@
 from http import HTTPStatus
+
+import matplotlib
 from flask import Flask, jsonify, redirect, url_for
 
 import src
+from src import config
+from src import controller
+from src import login_manager
+from src.api import API_BLUEPRINT
+from src.api.execution_plan_api import EXECUTION_PLAN_API_BLUEPRINT
 from src.encoders import CustomJSONEncoder
 from src.translations import gettext, pretty_date
-from src import config
-from src import login_manager
-from src import controller
-import matplotlib
 
 matplotlib.use("Agg")
 
@@ -23,7 +26,9 @@ app.register_blueprint(controller.VIEW_BLUEPRINT, url_prefix="/view")
 app.register_blueprint(controller.PUBLIC_VIEW_BLUEPRINT, url_prefix="/view")
 app.register_blueprint(controller.SCHEDULE_TASK_BLUEPRINT, url_prefix="/schedule_task")
 app.register_blueprint(controller.USER_BLUEPRINT, url_prefix="/user")
-app.register_blueprint(controller.API_BLUEPRINT, url_prefix="/api")
+
+API_BLUEPRINT.register_blueprint(EXECUTION_PLAN_API_BLUEPRINT)
+app.register_blueprint(API_BLUEPRINT)
 
 app.jinja_env.globals.update(gettext=gettext)
 app.jinja_env.globals.update(pretty_date=pretty_date)
@@ -39,6 +44,18 @@ def health_check():
 @app.errorhandler(HTTPStatus.NOT_FOUND)
 def page_not_found(e):
     return redirect(url_for("view_controller.home")), HTTPStatus.MOVED_PERMANENTLY
+
+
+@app.route('/list_routes')
+def list_routes():
+    routes = []
+    for rule in app.url_map.iter_rules():
+        routes.append({
+            'endpoint': rule.endpoint,
+            'methods': ','.join(rule.methods),
+            'url': str(rule),
+        })
+    return jsonify({'routes': routes})
 
 
 app.json_encoder = CustomJSONEncoder
