+++ src/controller/__init__.py
@@ -4,4 +4,3 @@ from src.controller.public_view_controller import PUBLIC_VIEW_BLUEPRINT
 from src.controller.execution_plan_controller import EXECUTION_PLAN_BLUEPRINT
 from src.controller.schedule_controller import SCHEDULE_TASK_BLUEPRINT
 from src.controller.user_controller import USER_BLUEPRINT
-from src.controller.api_controller import API_BLUEPRINT
